gestion double en cours

+++ libft/includes/libft.h
@@ -6,7 +6,7 @@
 /*   By: ebaudet <ebaudet@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2013/11/21 11:44:10 by ebaudet           #+#    #+#             */
-/*   Updated: 2019/02/05 16:57:12 by ebaudet          ###   ########.fr       */
+/*   Updated: 2019/02/06 00:37:50 by ebaudet          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -94,7 +94,7 @@ int		ft_isspace(int c);
 double	ft_atof(const char *str);
 int		ft_abs(int i);
 double	ft_fabs(double i);
-char	*ft_dtoa(double d, unsigned int precision);
+char	*ft_dtoa(long double d, unsigned int precision);
 int		ft_max(int a, int b);
 int		ft_min(int a, int b);
